% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{check_len}
\alias{check_len}
\title{Check if 'x$data' has length > 'len'}
\usage{
check_len(x, len)
}
\arguments{
\item{x}{list with data, result, and any errors already found}

\item{len}{minimum length 'x$data' can have}
}
\description{
Check if 'x$data' has length > 'len'
}

