% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{check_max_na}
\alias{check_max_na}
\title{Check if fraction of 'x$data' that is NA is <= 'max_na'}
\usage{
check_max_na(x, max_na, rm_na = FALSE)
}
\arguments{
\item{x}{list with data, result, and any errors already found}

\item{max_na}{maximum fraction of 'x$data' that can be NA}

\item{rm_na}{whether NAs should be removed once test is over}
}
\description{
Check if fraction of 'x$data' that is NA is <= 'max_na'
}

