% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{check_max_unq}
\alias{check_max_unq}
\title{Check if there aren't more than 'max_unq' classes in 'x$data'}
\usage{
check_max_unq(x, max_unq)
}
\arguments{
\item{x}{list with data, result, and any errors already found}

\item{max_unq}{maximum number of distinct classes 'x$data' can have}
}
\description{
Check if there aren't more than 'max_unq' classes in 'x$data'
}
