% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{check_min_unq}
\alias{check_min_unq}
\title{Check if there aren't less than 'min_unq' classes in 'x$data'}
\usage{
check_min_unq(x, min_unq)
}
\arguments{
\item{x}{list with data, result, and any errors already found}

\item{min_unq}{minimum number of distinct classes 'x$data' can have}
}
\description{
Check if there aren't less than 'min_unq' classes in 'x$data'
}

