% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare the profiles of two tables}
\usage{
compare(X, Y, ci = 0.05)
}
\arguments{
\item{X}{table used as standard for comparison}

\item{Y}{table to be evaluated}

\item{ci}{percentage to be used for the CI}
}
\description{
Compare the profiles of two tables
}
