% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.R
\name{diagnose}
\alias{diagnose}
\title{Obter inputs para uma bateria de testes}
\usage{
diagnose(data, exams)
}
\arguments{
\item{data}{banco de dados}

\item{exams}{testes a serem rodados}
}
\description{
Obter inputs para uma bateria de testes
}

