% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{dx}
\alias{dx}
\title{Diagnose model}
\usage{
dx(.model, ..., .out = TRUE, .viz = TRUE)
}
\arguments{
\item{.model}{A model.}

\item{...}{One or more unquoted dx functions (see details). User-defined
dx functions can receive arguments native to the function and will supersede
dx arguments. Leaving "..." blank will cause dx to automatically select
dx functions based on model class and family. See details for documentation
on selected dx_functions.}

\item{.out}{Print output to console. User-defined dx functions will overwrite
this argument.}

\item{.viz}{Print plots to plotting window. User-defined dx functions will
overwrite this argument.}
}
\description{
Diagnose model
}
\details{
A blank "..." will cause dx to automatically select dx functions
based on class and family (if applicable).

At this time only lm() class models are supported.

lm() diagnostics:
\itemize{
\item \code{\link{dx_cooks}}
\item \code{\link{dx_res_stand}}
\item \code{\link{dx_res_stud}}
\item \code{\link{dx_dfits}}
}
}
