% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_cooks.R
\name{dx_cooks}
\alias{dx_cooks}
\title{Cook's Distance}
\usage{
dx_cooks(.model, .threshold = 1, .out = TRUE, .viz = TRUE)
}
\arguments{
\item{.model}{A model}

\item{.threshold}{Cutoff point for significant values, used for output and visualisation.
Null argument will use a threshold of 1.}

\item{.out}{Print output to console.}

\item{.viz}{Print visualization to plotting window.}
}
\description{
Cook's Distance
}
