% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_dfits.R
\name{dx_dfits}
\alias{dx_dfits}
\title{DFITS}
\usage{
dx_dfits(.model, .threshold = 1, .out = TRUE, .viz = TRUE)
}
\arguments{
\item{.model}{A model}

\item{.threshold}{Cutoff point for significant values, used for output and visualization.
No argument will use a threshold of 1.}

\item{.out}{Print output to console.}

\item{.viz}{Print visualization to plotting window.}
}
\description{
DFITS
}
