% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_dif_fits.R
\name{dx_dif_fits}
\alias{dx_dif_fits}
\title{DFITS}
\usage{
dx_dif_fits(.model, .threshold = 1, .out = TRUE, .viz = TRUE)
}
\arguments{
\item{model}{A model}

\item{threshold}{Cutoff point for significant values, used for output and visualization.
No argument will use a threshold of 1.}

\item{out}{Print output to console.}

\item{visualisation}{Print plot to plotting window.}
}
\description{
DFITS
}
