% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_res_stand.R
\name{dx_res_stand}
\alias{dx_res_stand}
\title{Standardized Residuals}
\usage{
dx_res_stand(.model, .threshold = 3, .out = TRUE, .viz = TRUE)
}
\arguments{
\item{.model}{A model}

\item{.threshold}{Cutoff point for significant values, used for output and
visualization. A missing argument will use a threshold of 3.}

\item{.out}{Print output to console.}

\item{.visualisation}{Print plot to plotting window.}
}
\description{
Standardized Residuals
}
