% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_res_stud.R
\name{dx_res_stud}
\alias{dx_res_stud}
\title{Student Residuals}
\usage{
dx_res_stud(.model, .threshold = 2, .out = TRUE, .viz = TRUE)
}
\arguments{
\item{.model}{A model}

\item{.threshold}{Cutoff point for significant values, used for output and visualization.
No argument will use a threshold of 2.}

\item{.out}{Print output to console.}

\item{.viz}{Print visualization to plotting window.}
}
\description{
Student Residuals
}
