% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx_select.R
\name{dx_select}
\alias{dx_select}
\title{Select Model}
\usage{
dx_select(model)
}
\arguments{
\item{model}{A model.}
}
\description{
Internal function for dx, selects dx functions using model
class and family.
}
