% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine.R
\name{examine_}
\alias{examine_}
\title{Create summary statistics for every column in X (no grouping)}
\usage{
examine_(X)
}
\arguments{
\item{X}{table to be examined}
}
\description{
Create summary statistics for every column in X (no grouping)
}
