% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.R
\name{is_continuous}
\alias{is_continuous}
\alias{is_count}
\alias{is_money}
\alias{is_percentage}
\alias{is_quantity}
\title{Verificar se é um vetor de variável continua atendendo aos parâmetros especificados}
\usage{
is_continuous(x, min_val = -Inf, max_val = Inf, allow_na = TRUE,
  rational = TRUE, max_dec_places = Inf)

is_count(x, min_val = 0, max_val = Inf, allow_na = TRUE)

is_quantity(x, min_val = 0, max_val = Inf, allow_na = TRUE)

is_percentage(x, min_val = 0, max_val = 100, allow_na = TRUE)

is_money(x, min_val = 0, max_val = 10000, allow_na = TRUE,
  max_dec_places = 2)
}
\arguments{
\item{x}{vetor}

\item{min_val}{valor mínimo}

\item{max_val}{valor máximo}

\item{allow_na}{lógico indicando se o vetor pode ter NA's}

\item{rational}{lógico indicando se o vetor é de números racionais}

\item{max_dec_places}{número máximo de casas decimais}
}
\description{
Verificar se é um vetor de variável continua atendendo aos parâmetros especificados

Verificar se é um vetor de contagens (números, inteiros, [0, Inf])

Verificar se é um vetor de quantidades (números, racionais, [-Inf, Inf])

Verificar se é um vetor de porcentagens (números, racionais, [0, 100])

Verificar se é um vetor de valores monetários (números, racionais, [0, Inf])
}

