% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issues}
\alias{issues}
\title{Print messages with issues found in X (or in one of its columns)
  after running either \code{diagnose(X)} or \code{compare(X)}
  
This function interprets the results from \code{diagnose()} and
  \code{compare()}, and generates a comprehensive report printed as
  messages to the console; for more information about what gets printed
  consult \code{vignette("doctr_diagnose")} or \code{vignette("doctr_examine")}}
\usage{
issues(res, i = 0, verbose = FALSE)
}
\arguments{
\item{res}{A list returned from \code{diagnose()} or \code{compare()}}

\item{i}{Index or name of column from which to print issues}

\item{verbose}{Specify what issues were found in each column}
}
\description{
Print messages with issues found in X (or in one of its columns)
  after running either \code{diagnose(X)} or \code{compare(X)}
  
This function interprets the results from \code{diagnose()} and
  \code{compare()}, and generates a comprehensive report printed as
  messages to the console; for more information about what gets printed
  consult \code{vignette("doctr_diagnose")} or \code{vignette("doctr_examine")}
}
