% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.R
\name{print.logi_prob}
\alias{print.logi_prob}
\title{Imprimir um objeto logi_prob}
\usage{
\method{print}{logi_prob}(x, ...)
}
\arguments{
\item{x}{objeto logi_prob}

\item{...}{other arguments passed to print}
}
\description{
Imprimir um objeto logi_prob
}

