% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.R
\name{problems}
\alias{problems}
\title{Imprimir os problemas associados a uma coluna de uma tabela}
\usage{
problems(x, i = 0)
}
\arguments{
\item{x}{objeto logi_prob}

\item{i}{índice ou nome da coluna para a qual imprimir os problemas. i = 0 imprime 
todos os problemas}
}
\description{
Imprimir os problemas associados a uma coluna de uma tabela
}

