% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{problems_}
\alias{problems_}
\title{Create message with problems found in x}
\usage{
problems_(x, name)
}
\arguments{
\item{x}{list with data, result, and any errors already found}

\item{name}{name of column referring to 'x'}
}
\description{
Create message with problems found in x
}

