% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_chr}
\alias{profile_chr}
\title{Create profile for column of characters}
\usage{
profile_chr(x)
}
\arguments{
\item{x}{list with data of a column}
}
\description{
Create profile for column of characters
}
