% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{profile_dbl}
\alias{profile_dbl}
\title{Create profile for column of doubles}
\usage{
profile_dbl(x)
}
\arguments{
\item{x}{list with data of a column}
}
\description{
Create profile for column of doubles
}

