% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{profile_fct}
\alias{profile_fct}
\title{Create profile for column of characters}
\usage{
profile_fct(x)
}
\arguments{
\item{x}{list with data of a column}
}
\description{
Create profile for column of characters
}

