% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_num}
\alias{profile_num}
\title{Create profile for column of numerics}
\usage{
profile_num(x)
}
\arguments{
\item{x}{List with data of a column}
}
\description{
Create profile for column of numerics
}
