% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_txt}
\alias{profile_txt}
\title{Create profile for column of characters}
\usage{
profile_txt(x)
}
\arguments{
\item{x}{List with data of a column}
}
\description{
Create profile for column of characters
}
