% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_num}
\alias{report_num}
\title{Return summary of columns of numerics in a table}
\usage{
report_num(X, group = "")
}
\arguments{
\item{X}{list created by 'examine'}

\item{group}{group from which to retrieve summary}
}
\description{
Return summary of columns of numerics in a table
}
