% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.R
\name{run_tests}
\alias{run_tests}
\title{Rodar bateria de testes}
\usage{
run_tests(cols, funs, args)
}
\arguments{
\item{cols}{vetor de chr com nome das colunas}

\item{funs}{lista de funções aplicadas p/ cada coluna}

\item{args}{lista de argumentos passados p/ cada função}
}
\description{
Rodar bateria de testes
}

