% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{summary_dbl}
\alias{summary_dbl}
\title{Return summary of columns of doubles in a table}
\usage{
summary_dbl(X, group = "")
}
\arguments{
\item{X}{list created by 'examine'}

\item{group}{group from which to retrieve summary}
}
\description{
Return summary of columns of doubles in a table
}
