% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{translate}
\alias{translate}
\title{Convert vector of types to corresponding functions}
\usage{
translate(types)
}
\arguments{
\item{types}{vector with data types (money, count, etc.)}
}
\description{
Convert vector of types to corresponding functions
}
