% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_len}
\alias{check_len}
\title{Check if x$data has length > len}
\usage{
check_len(x, len)
}
\arguments{
\item{x}{List with data, result, and any errors already found}

\item{len}{Minimum length x$data can have}
}
\description{
Check if x$data has length > len
}
