% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_max_na}
\alias{check_max_na}
\title{Check if fraction of x$data that is NA is <= max_na}
\usage{
check_max_na(x, max_na, rm_na = FALSE)
}
\arguments{
\item{x}{List with data, result, and any errors already found}

\item{max_na}{Maximum fraction of x$data that can be NA}

\item{rm_na}{Whether NAs should be removed once test is over}
}
\description{
Check if fraction of x$data that is NA is <= max_na
}
