% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_max_unq}
\alias{check_max_unq}
\title{Check if there arent more than max_unq classes in x$data}
\usage{
check_max_unq(x, max_unq)
}
\arguments{
\item{x}{List with data, result, and any errors already found}

\item{max_unq}{Maximum number of distinct classes x$data can have}
}
\description{
Check if there arent more than max_unq classes in x$data
}
