% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mdp}
\alias{check_mdp}
\title{Check if no entry of x$data has more decimal places than mdp}
\usage{
check_mdp(x, mdp)
}
\arguments{
\item{x}{List with data, result, and any errors already found}

\item{mdp}{Maximum number of decimal places an entry in x$data can have}
}
\description{
Check if no entry of x$data has more decimal places than mdp
}
