% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_mfc}
\alias{check_mfc}
\title{Check if no classe represents more than mfc of x$data}
\usage{
check_mfc(x, mfc)
}
\arguments{
\item{x}{List with data, result, and any errors already found}

\item{mfc}{Maximum fraction of total for most frequent class}
}
\description{
Check if no classe represents more than mfc of x$data
}
