% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{compare}
\alias{compare}
\title{Compare the profiles of two tables}
\usage{
compare(X, Y)
}
\arguments{
\item{X}{table used as standard for comparison}

\item{Y}{table to be evaluated}
}
\description{
Compare the profiles of two tables
}

