% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{diagnose}
\alias{diagnose}
\title{Run tests on a table to check if it fits it's expected form}
\usage{
diagnose(X, exams)
}
\arguments{
\item{X}{table to run tests on}

\item{exams}{tests to be run on 'X'}
}
\description{
Run tests on a table to check if it fits it's expected form
}

