% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{examine}
\alias{examine}
\title{Create summary statistics for every column in 'X'}
\usage{
examine(X)
}
\arguments{
\item{X}{table to be examined}
}
\description{
Create summary statistics for every column in 'X'
}
