% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.R
\name{examine}
\alias{examine}
\title{An치lise explorat칩ria de um banco de dados}
\usage{
examine(data, cols)
}
\arguments{
\item{data}{banco de dados}

\item{cols}{vetor de chr com nome das colunas}
}
\description{
An치lise explorat칩ria de um banco de dados
}

