% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine.R
\name{examine}
\alias{examine}
\title{Create summary statistics for every column in 'X'}
\usage{
examine(X, group = 0)
}
\arguments{
\item{X}{table to be examined}

\item{group}{variable to group X by before examining}
}
\description{
Create summary statistics for every column in 'X'
}
