% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{guess_exams}
\alias{guess_exams}
\title{Try to guess which exam a dataset should go through}
\usage{
guess_exams(X)
}
\arguments{
\item{X}{table to be examined}
}
\description{
Try to guess which exam a dataset should go through
}

