% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issues_}
\alias{issues_}
\title{Create message with issues found in x}
\usage{
issues_(x, name, verbose)
}
\arguments{
\item{x}{List with data, result, and any errors already found}

\item{name}{Name of column referring to x}

\item{verbose}{Specify the issues found in each column}
}
\description{
Create message with issues found in x
}
