% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{problems}
\alias{problems}
\title{Print messages with problems found in 'X' (or in one of its columns)}
\usage{
problems(X, i = 0)
}
\arguments{
\item{X}{list of lists with data, result, and any errors already found for each column}

\item{i}{index of column from which to print problems}
}
\description{
Print messages with problems found in 'X' (or in one of its columns)
}

