% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctr.R
\name{summary_chr}
\alias{summary_chr}
\title{Return summary of columns of characters in a table}
\usage{
summary_chr(X)
}
\arguments{
\item{X}{list created by 'examine'}
}
\description{
Return summary of columns of characters in a table
}

